Update based on eslint
