+++ .eslintrc.js
@@ -15,6 +15,7 @@ module.exports = {
     sourceType: 'module',
   },
   plugins: [
+    'only-warn',
     'react',
   ],
   rules: {
