+++ src/About.jsx
@@ -1,5 +1,7 @@
+import React from 'react';
+
 function About() {
   return <h2>About this</h2>;
 }
 
-export default About
+export default About;
