+++ src/App.jsx
@@ -1,0 +1,45 @@
+import React from 'react';
+import './App.scss';
+import { withStyles } from '@material-ui/core/styles';
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+} from 'react-router-dom';
+import About from './About';
+import Search from './Search';
+import Home from './Home';
+
+const styles = (theme) => ({
+  root: {
+    flexGrow: 1,
+  },
+  menuButton: {
+    marginRight: theme.spacing(2),
+  },
+  title: {
+    flexGrow: 1,
+  },
+});
+function App() {
+  return (
+    <>
+      <Router>
+        <Switch>
+          <Route path="/about">
+            <About />
+          </Route>
+          <Route path="/search">
+            <Search />
+          </Route>
+          <Route path="/">
+            <Home />
+          </Route>
+        </Switch>
+      </Router>
+    </>
+  );
+}
+
+// export default App
+export default withStyles(styles)(App);
