+++ src/index.jsx
@@ -1,15 +1,16 @@
+import 'typeface-roboto';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
+import { HashRouter } from 'react-router-dom';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import { HashRouter } from "react-router-dom"
 
 ReactDOM.render(
   <React.StrictMode>
     <HashRouter><App /></HashRouter>
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById('root'),
 );
 
 // If you want to start measuring performance in your app, pass a function
